backend/requirements-dev.txt
-r requirements.txt

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.2.0

# Code Quality
flake8==7.0.0
black==23.12.1
isort==5.13.2
mypy==1.8.0
django-stubs==4.2.7

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.20.0
watchdog==3.0.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.4